open: the recomposition carries each pair once, and only a layer that derives geometry accumulates it
The referee's findings on the three commits below.

A growth on a form that holds no rows offers the artifact's whole membership as the pairs
the column gains — it has no held set to subtract one from — and the recomposition
concatenated that with the column's own pairs. A list row could then name one ordinal
twice, which counts the artifact twice in the histogram, twice in the declared size the
proportional criterion divides by, and twice in the accumulated centroid. The feed now
makes the same subtraction `amend` makes: a pair the column already carries is skipped
and the rest is deduped. The new differential grows an artifact in the tick that
recomposes the column and compares every answer, including each artifact's declared size
against the twin's membership cardinality; without the fix the grown artifact's masked
count is 1,000 where it should be 500.

The accumulation is built only where the layer derives a centroid or a box. A count-only
layer paid a grid position per visible row and 36 B an ordinal for a number no response
carried.

Both routes to a centroid now sum in u64 rather than f64. A row space is u32-addressed
and a coordinate is a u32, so a per-axis sum is at most (2^32 - 1)^2 and fits exactly;
the chunked reduction and the sequential walk therefore agree bit for bit, which
floating-point addition past 2^53 would not give. The f64 sum's own justification — that
a u64 would overflow past 2^32 members — was wrong: a membership cannot exceed the row
count.

Beside those: the 36 B entry is priced where the 4 B figure is quoted, with the reason
the default cache bound does not move and the note that a viewport and a browse page
build two entries for one level by design; the accumulation's transient is stated as per
chunk in flight, per build, per session; visible_rows prices the copy it makes of the
span's visible rows and says why MaskedSet has no iterator to avoid it; and a stale
citation of hold_rows is replaced by what the build route actually conditions on.

+++ crates/tessera-engine/src/derived.rs
@@ -241,7 +241,10 @@ impl<'a> RowLocator<'a> {
 /// **The same two quantities by another route, not two definitions of them.** A centroid is the
 /// mean of `membership ∩ M_auth`'s positions and a box their extremes; summing and comparing as the
 /// rows are read gives the same numbers as materialising the rows and traversing them, which is
-/// what `tests/artifact_row_major.rs` asserts against the artifact-major route.
+/// what `tests/artifact_row_major.rs` asserts against the artifact-major route. **Exactly the same
+/// numbers**: both sum in `u64` — the sum cannot overflow one and is exact in one — so the chunked
+/// reduction and the sequential walk agree bit for bit, which floating-point addition past 2^53
+/// would not give.
 ///
 /// ⊘ **A hull is not here**, and cannot be: it is a function of the positions themselves rather
 /// than an accumulation over them. A layer deriving one keeps the artifact-major form
@@ -294,13 +297,16 @@ pub fn compute(
         }
     }
 
-    // Summed as `f64` rather than `u64`: the grid is 2^32 wide, so a membership past ~2^32 members
-    // would overflow a `u64` sum, and the mean is fractional in any case.
-    let (mut sx, mut sy) = (0.0f64, 0.0f64);
+    // **Summed as `u64`, which is exact and cannot overflow**: a membership is a set of rows, a
+    // row space is `u32`-addressed and a coordinate is a `u32`, so a per-axis sum is at most
+    // `(2^32 - 1)^2`. The mean is fractional and the division is the only floating-point step, so
+    // this and `crate::histogram::MaskedGeometry`'s accumulation produce the same number rather
+    // than two roundings of it.
+    let (mut sx, mut sy) = (0u64, 0u64);
     let mut b = [u32::MAX, u32::MAX, 0u32, 0u32];
     for p in &positions {
-        sx += p[0] as f64;
-        sy += p[1] as f64;
+        sx += u64::from(p[0]);
+        sy += u64::from(p[1]);
         b[0] = b[0].min(p[0]);
         b[1] = b[1].min(p[1]);
         b[2] = b[2].max(p[0]);
@@ -309,7 +315,7 @@ pub fn compute(
 
     if want_centroid {
         let n = positions.len() as f64;
-        out.centroid = Some([sx / n, sy / n]);
+        out.centroid = Some([sx as f64 / n, sy as f64 / n]);
     }
     if want_box {
         out.bbox = Some(b);
